% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plot_theme.R
\name{new_plot_theme}
\alias{new_plot_theme}
\title{Create a customized PKPDsim plot theme}
\usage{
new_plot_theme(update = NULL)
}
\arguments{
\item{update}{list containing the plot elements to be updated. Run `new_plot_theme()` with no arguments to show an overview of available plot elements.}
}
\value{
A list with vpc theme specifiers
}
\description{
Create a customized PKPDsim plot theme
}
