% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PKPDsim_data.R
\name{plot.PKPDsim_x}
\alias{plot.PKPDsim_x}
\title{Generic plotting function for PKPDsim objects}
\usage{
\method{plot}{PKPDsim_x}(x, only_obs = TRUE, obs_x = NULL, regimen = NULL,
  lines = NULL, show_single = list(obs = TRUE, spaghetti = TRUE, ci = FALSE,
  median = FALSE, regimen = TRUE), show_population = list(obs = FALSE,
  spaghetti = TRUE, ci = FALSE, median = TRUE, regimen = TRUE), ci = c(0.05,
  0.95), theme = NULL, xlim = NULL, ylim = NULL, width = 6,
  height = 4, legend = NULL, show_series = NULL, show = NULL,
  log_y = FALSE, target = NULL, target_as_ribbon = FALSE,
  scale_colour_values = NULL, scale_linetype_values = NULL,
  labels = list(x = "Time (hours)", y = "Concentration (mg/L)"),
  return_svg = FALSE, ...)
}
\arguments{
\item{x}{PKPDsim-simulated x}

\item{only_obs}{Only plot observation compartment}

\item{obs_x}{Also add observation (e.g. TDM) x}

\item{regimen}{regimen specification from PKPDsim::new_regimen()}

\item{lines}{show lines?}

\item{show_single}{definition for plots of single patients}

\item{show_population}{definition for plots of populations}

\item{ci}{confidence interval, specified as vector of two values between 0 and 1, default is `c(0.05, 0.95)`.}

\item{theme}{theme}

\item{xlim}{xlim passed to ggplot}

\item{ylim}{ylim passed to ggplot}

\item{width}{width of plot if saved}

\item{height}{height of plot if saved}

\item{legend}{list of labels that override defaults, e.g. list("individual" = "ind.pred")}

\item{show_series}{show only specified types / series}

\item{show}{definition of what to show in plot, overrides `show_single` and `show_population`. NULL by default}

\item{log_y}{log y axis?}

\item{target}{vector of target values (dependent variable), will be shown as vertical lines}

\item{target_as_ribbon}{show target (only when vector of 2) as ribbon (TRUE), or just as two lines (FALSE)}

\item{scale_colour_values}{values for colour scale}

\item{scale_linetype_values}{values for linetype scale}

\item{labels}{list with `x` and `y` labels}

\item{return_svg}{return the svg plot definition instead of ggplot2 object. FALSE by default}

\item{...}{rest}
}
\description{
Generic plotting function for PKPDsim objects
}
