% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_overlay.R
\name{png_overlay}
\alias{png_overlay}
\title{Generate an overlay for ggplot2 plots from a PNG file}
\usage{
png_overlay(png = NULL, x_sel = c(0, 1), y_sel = c(0, 1), x_range = c(0,
  24), y_range = c(0, 100))
}
\arguments{
\item{png}{PNG file}

\item{x_sel}{relative x-viewwindow for PNG, default `c(0,1)`}

\item{y_sel}{relative y-viewwindow for PNG, default `c(0,1)`}

\item{x_range}{x-range on scale of ggplot to show / stretch the PNG}

\item{y_range}{y-range on scale of ggplot to show / stretch the PNG}
}
\description{
Generate an overlay for ggplot2 plots from a PNG file
}
