% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.PKPDsim_data.R
\name{plot.PKPDsim_data}
\alias{plot.PKPDsim_data}
\title{Generic plotting function for PKPDsim objects}
\usage{
\method{plot}{PKPDsim_data}(data, only_obs = TRUE, tooltips = FALSE,
  show_single = list(obs = TRUE, spaghetti = TRUE, ci = FALSE, median = FALSE,
  regimen = TRUE), show_population = list(obs = FALSE, spaghetti = TRUE, ci =
  FALSE, median = TRUE, regimen = TRUE), show = NULL, log_y = FALSE,
  target = NULL, target_as_ribbon = TRUE, labels = list(x =
  "Time (hours)", y = "Concentration (mg/L)"), return_svg = FALSE, ...)
}
\arguments{
\item{data}{PKPDsim-simulated data}

\item{only_obs}{Only plot observation compartment}

\item{tooltips}{}

\item{show_single}{definition for plots of single patients}

\item{show_population}{definition for plots of populations}

\item{show}{definition of what to show in plot, overrides `show_single` and `show_population`. NULL by default}

\item{target}{vector of target values (dependent variable), will be shown as vertical lines}

\item{target_as_ribbon}{show target (only when vector of 2) as ribbon (TRUE), or just as two lines (FALSE)}

\item{labels}{list with `x` and `y` labels}

\item{return_svg}{return the svg plot definition instead of ggplot2 object. FALSE by default}

\item{...}{rest}
}
\description{
Generic plotting function for PKPDsim objects
}

