% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PKPDsim_data.R
\name{plot.PKPDsim_data}
\alias{plot.PKPDsim_data}
\title{Generic plotting function for PKPDsim objects}
\usage{
\method{plot}{PKPDsim_data}(data, only_obs = TRUE, obs_data = NULL,
  regimen = NULL, lines = NULL, show_single = list(obs = TRUE, spaghetti =
  TRUE, ci = FALSE, median = FALSE, regimen = TRUE),
  show_population = list(obs = FALSE, spaghetti = TRUE, ci = FALSE, median =
  TRUE, regimen = TRUE), theme = NULL, xlim = NULL, ylim = NULL,
  width = 6, height = 4, legend = NULL, show_series = NULL,
  show = NULL, log_y = FALSE, target = NULL, target_as_ribbon = FALSE,
  scale_colour_values = NULL, scale_linetype_values = NULL,
  labels = list(x = "Time (hours)", y = "Concentration (mg/L)"),
  return_svg = FALSE, ...)
}
\arguments{
\item{data}{PKPDsim-simulated data}

\item{only_obs}{Only plot observation compartment}

\item{obs_data}{Also add observation (e.g. TDM) data}

\item{regimen}{regimen specification from PKPDsim::new_regimen()}

\item{show_single}{definition for plots of single patients}

\item{show_population}{definition for plots of populations}

\item{legend}{list of labels that override defaults, e.g. list("individual" = "ind.pred")}

\item{show}{definition of what to show in plot, overrides `show_single` and `show_population`. NULL by default}

\item{target}{vector of target values (dependent variable), will be shown as vertical lines}

\item{target_as_ribbon}{show target (only when vector of 2) as ribbon (TRUE), or just as two lines (FALSE)}

\item{scale_colour_values}{values for colour scale}

\item{scale_linetype_values}{values for linetype scale}

\item{labels}{list with `x` and `y` labels}

\item{return_svg}{return the svg plot definition instead of ggplot2 object. FALSE by default}

\item{...}{rest}
}
\description{
Generic plotting function for PKPDsim objects
}

